little fixes and AirShare callback hookup

+++ sdk/src/main/java/pro/dbro/airshare/session/SessionMessage.java
@@ -111,7 +111,7 @@ public abstract class SessionMessage {
     /**
      * @return the length of the blob body in bytes
      */
-    public long getBodyLengthBytes() {
+    public int getBodyLengthBytes() {
         return bodyLengthBytes;
     }
